+++ system-tests/projects/no-specs-vue/src/components/TheWelcome.vue
@@ -12,10 +12,16 @@ import SupportIcon from './icons/IconSupport.vue'
     <template #icon>
       <DocumentationIcon />
     </template>
-    <template #heading>Documentation</template>
+    <template #heading>
+      Documentation
+    </template>
 
     Vue’s
-    <a href="https://vuejs.org/" target="_blank" rel="noopener">official documentation</a>
+    <a
+      href="https://vuejs.org/"
+      target="_blank"
+      rel="noopener"
+    >official documentation</a>
     provides you with all information you need to get started.
   </WelcomeItem>
 
@@ -23,20 +29,40 @@ import SupportIcon from './icons/IconSupport.vue'
     <template #icon>
       <ToolingIcon />
     </template>
-    <template #heading>Tooling</template>
+    <template #heading>
+      Tooling
+    </template>
 
     This project is served and bundled with
-    <a href="https://vite.dev/guide/features.html" target="_blank" rel="noopener">Vite</a>. The
+    <a
+      href="https://vite.dev/guide/features.html"
+      target="_blank"
+      rel="noopener"
+    >Vite</a>. The
     recommended IDE setup is
-    <a href="https://code.visualstudio.com/" target="_blank" rel="noopener">VSCode</a> +
-    <a href="https://github.com/johnsoncodehk/volar" target="_blank" rel="noopener">Volar</a>. If
+    <a
+      href="https://code.visualstudio.com/"
+      target="_blank"
+      rel="noopener"
+    >VSCode</a> +
+    <a
+      href="https://github.com/johnsoncodehk/volar"
+      target="_blank"
+      rel="noopener"
+    >Volar</a>. If
     you need to test your components and web pages, check out
-    <a href="https://www.cypress.io/" target="_blank" rel="noopener">Cypress</a> and
-    <a href="https://on.cypress.io/component" target="_blank" rel="noopener"
-      >Cypress Component Testing</a
-    >.
+    <a
+      href="https://www.cypress.io/"
+      target="_blank"
+      rel="noopener"
+    >Cypress</a> and
+    <a
+      href="https://on.cypress.io/component"
+      target="_blank"
+      rel="noopener"
+    >Cypress Component Testing</a>.
 
-    <br />
+    <br>
 
     More instructions are available in <code>README.md</code>.
   </WelcomeItem>
@@ -45,15 +71,37 @@ import SupportIcon from './icons/IconSupport.vue'
     <template #icon>
       <EcosystemIcon />
     </template>
-    <template #heading>Ecosystem</template>
+    <template #heading>
+      Ecosystem
+    </template>
 
     Get official tools and libraries for your project:
-    <a href="https://pinia.vuejs.org/" target="_blank" rel="noopener">Pinia</a>,
-    <a href="https://router.vuejs.org/" target="_blank" rel="noopener">Vue Router</a>,
-    <a href="https://test-utils.vuejs.org/" target="_blank" rel="noopener">Vue Test Utils</a>, and
-    <a href="https://github.com/vuejs/devtools" target="_blank" rel="noopener">Vue Dev Tools</a>. If
+    <a
+      href="https://pinia.vuejs.org/"
+      target="_blank"
+      rel="noopener"
+    >Pinia</a>,
+    <a
+      href="https://router.vuejs.org/"
+      target="_blank"
+      rel="noopener"
+    >Vue Router</a>,
+    <a
+      href="https://test-utils.vuejs.org/"
+      target="_blank"
+      rel="noopener"
+    >Vue Test Utils</a>, and
+    <a
+      href="https://github.com/vuejs/devtools"
+      target="_blank"
+      rel="noopener"
+    >Vue Dev Tools</a>. If
     you need more resources, we suggest paying
-    <a href="https://github.com/vuejs/awesome-vue" target="_blank" rel="noopener">Awesome Vue</a>
+    <a
+      href="https://github.com/vuejs/awesome-vue"
+      target="_blank"
+      rel="noopener"
+    >Awesome Vue</a>
     a visit.
   </WelcomeItem>
 
@@ -61,17 +109,33 @@ import SupportIcon from './icons/IconSupport.vue'
     <template #icon>
       <CommunityIcon />
     </template>
-    <template #heading>Community</template>
+    <template #heading>
+      Community
+    </template>
 
     Got stuck? Ask your question on
-    <a href="https://chat.vuejs.org" target="_blank" rel="noopener">Vue Land</a>, our official
+    <a
+      href="https://chat.vuejs.org"
+      target="_blank"
+      rel="noopener"
+    >Vue Land</a>, our official
     Discord server, or
-    <a href="https://stackoverflow.com/questions/tagged/vue.js" target="_blank" rel="noopener"
-      >StackOverflow</a
-    >. You should also subscribe to
-    <a href="https://news.vuejs.org" target="_blank" rel="noopener">our mailing list</a> and follow
+    <a
+      href="https://stackoverflow.com/questions/tagged/vue.js"
+      target="_blank"
+      rel="noopener"
+    >StackOverflow</a>. You should also subscribe to
+    <a
+      href="https://news.vuejs.org"
+      target="_blank"
+      rel="noopener"
+    >our mailing list</a> and follow
     the official
-    <a href="https://twitter.com/vuejs" target="_blank" rel="noopener">@vuejs</a>
+    <a
+      href="https://twitter.com/vuejs"
+      target="_blank"
+      rel="noopener"
+    >@vuejs</a>
     twitter account for latest news in the Vue world.
   </WelcomeItem>
 
@@ -79,10 +143,16 @@ import SupportIcon from './icons/IconSupport.vue'
     <template #icon>
       <SupportIcon />
     </template>
-    <template #heading>Support Vue</template>
+    <template #heading>
+      Support Vue
+    </template>
 
     As an independent project, Vue relies on community backing for its sustainability. You can help
     us by
-    <a href="https://vuejs.org/sponsor/" target="_blank" rel="noopener">becoming a sponsor</a>.
+    <a
+      href="https://vuejs.org/sponsor/"
+      target="_blank"
+      rel="noopener"
+    >becoming a sponsor</a>.
   </WelcomeItem>
 </template>