+++ cli/CHANGELOG.md
@@ -6,7 +6,7 @@ _Released 1/28/2025 (PENDING)_
 **Bugfixes:**
 
 - Fixed an issue where Cypress would incorrectly navigate to `about:blank` when test isolation was disabled and the last test would fail and then retry. Fixes [#28527](https://github.com/cypress-io/cypress/issues/28527).
-- Fixed an issue where non-HTMLElement(s) may failing during assertions. Fixes [#30944](https://github.com/cypress-io/cypress/issues/30944)
+- Fixed an issue where non-HTMLElement(s) may fail during assertions. Fixes [#30944](https://github.com/cypress-io/cypress/issues/30944)
 
 ## 14.0.0